Move LiteHierarchicaleModelDefinition together with EnergyBalanceHierarchicalModel and tidy.
