+++ src/HealthGPS/repository.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "disease_definition.h"
+#include "energy_balance_hierarchical_model.h"
 #include "energy_balance_model.h"
 #include "hierarchical_model_types.h"
 #include "interfaces.h"
